requirements_windows.txt
absl-py==0.9.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
commentjson==0.8.2
cycler==0.10.0
decorator==4.4.1
google-auth==1.10.2
google-auth-oauthlib==0.4.1
grpcio==1.26.0
idna==2.8
imageio==2.6.1
kiwisolver==1.1.0
lark-parser==0.8.1
Markdown==3.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scipy==1.4.1
six==1.14.0
tensorboard==2.1.0
https://download.pytorch.org/whl/cu101/torch-1.4.0-cp37-cp37m-win_amd64.whl
https://download.pytorch.org/whl/cu101/torchvision-0.5.0-cp37-cp37m-win_amd64.whl
urllib3==1.25.8
Werkzeug==0.16.0
wincertstore==0.2

